Create Text from image.py
The function TextFromImage(path) inputs the path of an image (path) and outputs the text extracted from the image.
The function requires pytesseract and PIL. An executable must be run before the function is used. The executable needs to be run once at the beginning of the script, after which any number of images can be processed.